// src/app/state/todo/todo.action-types.ts
export enum TODO_ACTIONS_TYPES {
  GET_TODOS = '[GET_TODOS]',
  GET_TODOS_SUCCESS = '[GET_TODOS] SUCCESS',
  GET_TODOS_FAILURE = '[GET_TODOS] FAILURE',

  ADD_TODO = '[ADD_TODO]',
  ADD_TODO_SUCCESS = '[ADD_TODO] SUCCESS',
  ADD_TODO_FAILURE = '[ADD_TODO] FAILURE',

  DELETE_TODO = '[DELETE_TODO]',
  DELETE_TODO_SUCCESS = '[DELETE_TODO] SUCCESS',
  DELETE_TODO_FAILURE = '[DELETE_TODO] FAILURE',
}